requirements.txt
pandas
mysql-connector-python==8.0.28
loguru